Added cmd download data with description in README
To registration.MyUser model added filed rated_books.
In registration was made a new model RatedBooks.
It was made in order to give the user a way of marking books to be
read in some future.

Migrations added.
README was enlarged by description of how to put the data from .csv
files to database.